Move TestPrintableGraph to printer_test.py
